requirements.txt
llama-cpp-python>=0.2.6
typing-extensions>=4.5.0
numpy>=1.24.0 